requirements.txt
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
idna==3.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
pyTelegramBotAPI==4.13.0
python-decouple==3.8
pytube==15.0.0
requests==2.31.0
service-identity==23.1.0
telebot==0.0.5
urllib3==2.0.4
